# packages/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: EcoLLM API
  version: "0.1.0"
  description: |
    Carbon-aware LLM inference platform.
    All requests require a Bearer API key unless noted otherwise.
  contact:
    email: api@ecollm.io
servers:
  - url: https://api.ecollm.io
    description: Production
  - url: https://staging-api.ecollm.io
    description: Staging
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /v1/usage:
    get:
      summary: Get usage aggregates for the current org
      operationId: getUsage
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [hourly, daily, monthly]
            default: daily
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Usage aggregate list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: eco-sk-...
...
  schemas:
...
    UsageResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object